requirements.txt
Flask
psycopg2-binary
python-dotenv
Flask-Cors
PyJWT>=2.8.0
                                                                                                                                                                                                                                                                                    gunicorn
